introduce ModuleMetrics interface add unit test for ModuleMetrics
