+++ pkg/module/sync/runner_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -14,6 +15,12 @@ import (
 	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
 	"github.com/kyma-project/lifecycle-manager/pkg/module/sync"
 	"github.com/kyma-project/lifecycle-manager/pkg/testutils"
+)
+
+const (
+	InvalidModulePrefix = "invalid_"
+	ModuleShouldKeep    = "ModuleShouldKeep"
+	ModuleToBeRemoved   = "ModuleToBeRemoved"
 )
 
 func moduleDeletedSuccessfullyMock(_ context.Context, _ client.Object) error {
@@ -24,12 +31,69 @@ func moduleStillExistsInClusterMock(_ context.Context, _ client.Object) error {
 	return apierrors.NewAlreadyExists(schema.GroupResource{}, "module-still-exists")
 }
 
-//nolint:funlen
+type ModuleMockMetrics struct {
+	mock.Mock
+}
+
+func (m *ModuleMockMetrics) RemoveModuleStateMetrics(kymaName, moduleName string) {
+	m.Called(kymaName, moduleName)
+}
+
+func TestMetricsOnDeleteNoLongerExistingModuleStatus(t *testing.T) {
+	t.Parallel()
+	tests := []struct {
+		name                         string
+		ModuleInStatus               string
+		getModule                    sync.GetModuleFunc
+		expectModuleMetricsGetCalled bool
+	}{
+		{
+			"When status.modules contains Manifest not found in cluster, expect RemoveModuleStateMetrics get called",
+			ModuleToBeRemoved,
+			moduleDeletedSuccessfullyMock,
+			true,
+		},
+		{
+			"When status.modules contains Manifest still exits in cluster, expect RemoveModuleStateMetrics not called",
+			ModuleToBeRemoved,
+			moduleStillExistsInClusterMock,
+			false,
+		},
+		{
+			"When status.modules contains not valid Manifest, expect RemoveModuleStateMetrics get called",
+			InvalidModulePrefix + ModuleToBeRemoved,
+			moduleStillExistsInClusterMock,
+			true,
+		},
+		{
+			"When status.modules contains module in spec.module, expect RemoveModuleStateMetrics not called",
+			ModuleShouldKeep,
+			moduleStillExistsInClusterMock,
+			false,
+		},
+	}
+
+	for _, testCase := range tests {
+		testCase := testCase
+		t.Run(testCase.name, func(t *testing.T) {
+			t.Parallel()
+			kyma := testutils.NewTestKyma("test-kyma")
+			configureModuleInKyma(kyma, []string{ModuleShouldKeep}, []string{testCase.ModuleInStatus})
+			mockMetrics := &ModuleMockMetrics{}
+			const methodToBeCalled = "RemoveModuleStateMetrics"
+			mockMetrics.On(methodToBeCalled, kyma.Name, testCase.ModuleInStatus).Return()
+			sync.DeleteNoLongerExistingModuleStatus(context.TODO(), kyma, testCase.getModule, mockMetrics)
+			if testCase.expectModuleMetricsGetCalled {
+				mockMetrics.AssertCalled(t, methodToBeCalled, kyma.Name, testCase.ModuleInStatus)
+			} else {
+				mockMetrics.AssertNotCalled(t, methodToBeCalled, kyma.Name, testCase.ModuleInStatus)
+			}
+		})
+	}
+}
+
 func TestDeleteNoLongerExistingModuleStatus(t *testing.T) {
 	t.Parallel()
-	const InvalidModulePrefix = "invalid_"
-	const ModuleShouldKeep = "ModuleShouldKeep"
-	const ModuleToBeRemoved = "ModuleToBeRemoved"
 	tests := []struct {
 		name                        string
 		ModulesInKymaSpec           []string
@@ -44,7 +108,6 @@ func TestDeleteNoLongerExistingModuleStatus(t *testing.T) {
 			[]string{ModuleShouldKeep},
 			moduleDeletedSuccessfullyMock,
 		},
-
 		{
 			"When status.modules contains invalid modules not in spec.module, expect removed and spec.module keep",
 			[]string{ModuleShouldKeep},
@@ -88,23 +151,7 @@ func TestDeleteNoLongerExistingModuleStatus(t *testing.T) {
 		t.Run(testCase.name, func(t *testing.T) {
 			t.Parallel()
 			kyma := testutils.NewTestKyma("test-kyma")
-			for _, moduleName := range testCase.ModulesInKymaSpec {
-				module := v1beta2.Module{
-					Name: moduleName,
-				}
-				kyma.Spec.Modules = append(kyma.Spec.Modules, module)
-			}
-			for _, moduleName := range testCase.ModulesInKymaStatus {
-				manifest := &v1beta2.TrackingObject{}
-				if strings.Contains(moduleName, InvalidModulePrefix) {
-					manifest = nil
-				}
-				module := v1beta2.ModuleStatus{
-					Name:     moduleName,
-					Manifest: manifest,
-				}
-				kyma.Status.Modules = append(kyma.Status.Modules, module)
-			}
+			configureModuleInKyma(kyma, testCase.ModulesInKymaSpec, testCase.ModulesInKymaStatus)
 			sync.DeleteNoLongerExistingModuleStatus(context.TODO(), kyma, testCase.getModule, nil)
 			var modulesInFinalModuleStatus []string
 			for _, moduleStatus := range kyma.Status.Modules {
@@ -114,5 +161,28 @@ func TestDeleteNoLongerExistingModuleStatus(t *testing.T) {
 			sort.Strings(modulesInFinalModuleStatus)
 			assert.Equal(t, testCase.ExpectedModulesInKymaStatus, modulesInFinalModuleStatus)
 		})
+	}
+}
+
+func configureModuleInKyma(
+	kyma *v1beta2.Kyma,
+	modulesInKymaSpec, modulesInKymaStatus []string,
+) {
+	for _, moduleName := range modulesInKymaSpec {
+		module := v1beta2.Module{
+			Name: moduleName,
+		}
+		kyma.Spec.Modules = append(kyma.Spec.Modules, module)
+	}
+	for _, moduleName := range modulesInKymaStatus {
+		manifest := &v1beta2.TrackingObject{}
+		if strings.Contains(moduleName, InvalidModulePrefix) {
+			manifest = nil
+		}
+		module := v1beta2.ModuleStatus{
+			Name:     moduleName,
+			Manifest: manifest,
+		}
+		kyma.Status.Modules = append(kyma.Status.Modules, module)
 	}
 }
